refactor: update import paths for mock models in tests
