+++ src/stories/shared-mock-repositories/info/DataverseInfoMockRepository.ts
@@ -5,9 +5,9 @@ import { TermsOfUse } from '@/info/domain/models/TermsOfUse'
 import { DataverseInfoRepository } from '@/info/domain/repositories/DataverseInfoRepository'
 import { Setting } from '@/settings/domain/models/Setting'
 import { ZipDownloadLimit } from '@/settings/domain/models/ZipDownloadLimit'
-import { DatasetMetadataExportFormatsMother } from '@tests/component/info/models/DatasetMetadataExportFormatsMother'
-import { DataverseVersionMother } from '@tests/component/info/models/DataverseVersionMother'
-import { TermsOfUseMother } from '@tests/component/info/models/TermsOfUseMother'
+import { DatasetMetadataExportFormatsMother } from '@tests/component/info/domain/models/DatasetMetadataExportFormatsMother'
+import { DataverseVersionMother } from '@tests/component/info/domain/models/DataverseVersionMother'
+import { TermsOfUseMother } from '@tests/component/info/domain/models/TermsOfUseMother'
 import { SettingMother } from '@tests/component/settings/domain/models/SettingMother'
 import { FakerHelper } from '@tests/component/shared/FakerHelper'
